chore(worker): move worker exports into separate files
Closes #91
